requirements.txt
Flask==2.2.5
flask-cors==4.0.0
pandas==1.5.3
tqdm==4.65.0
qdrant-client==1.6.0
google-generativeai==0.3.1
scikit-learn==1.2.2
python-dotenv==1.0.0
openpyxl==3.1.2